pyproject.toml
[tool.poetry]
name = "textguard-ai"
version = "1.0.0"
description = "Advanced ML/NLP service for text toxicity classification with enterprise architecture"
authors = ["TextGuardAI Team"]
readme = "README.md"
packages = [{include = "app"}, {include = "domain"}, {include = "infrastructure"}, {include = "application"}]

[tool.poetry.dependencies]
python = "^3.10"
# ML/AI
torch = "^2.1.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
nltk = "^3.8.1"
tqdm = "^4.66.0"

# Web Framework
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
gunicorn = "^21.2.0"

# Database
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"

# Validation & Serialization
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"

# Dependency Injection
dependency-injector = "^4.41.0"

# Background Tasks
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"

# HTTP Clients
httpx = "^0.25.0"

# Configuration & Environment
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"

# Logging
structlog = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

